requirements.txt
flask
flask_login
bcrypt
mysql-connector-python